+++ Services/LocalTokenUsageService.cs
@@ -1,9 +1,9 @@
 using System.IO;
 using System.Text;
 using System.Text.Json;
-using CodexQuotaFloat.Models;
+using UsageLens.Models;
 
-namespace CodexQuotaFloat.Services;
+namespace UsageLens.Services;
 
 public sealed class LocalTokenUsageService
 {